+++ learn-ucore/kern/net/tcpip/src/tcp/tcpsndlen.c
@@ -27,12 +27,12 @@ int tcpsndlen(struct tcb *ptcb, bool rexmt, unsigned int *poff)
 	else
 	{
 		//没有被确认的数据量
-		*poff = ptcb->tcb_snext - ptcb->tcb_suna;
+		*poff = (unsigned int)(ptcb->tcb_snext - ptcb->tcb_suna);
 	}
 	//发送缓冲区中个可以发出去的数据量
 	datalen = ptcb->tcb_sbcount - *poff;
 	//发送的数据量不能超过对端的通告窗口
-	datalen = min(datalen, ptcb->tcb_swindow);
+	datalen = (unsigned int)min(datalen, ptcb->tcb_swindow);
 
 	//同样，发送的数据量，不能超过对端的max segment size
 	return min(datalen, ptcb->tcb_smss);
